+++ module/Core/config/event_dispatcher.config.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Core;
 
 use Laminas\ServiceManager\AbstractFactory\ConfigAbstractFactory;
-use PhpAmqpLib\Connection\AMQPStreamConnection;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Shlinkio\Shlink\CLI\Util\GeolocationDbUpdater;
+use Shlinkio\Shlink\Common\RabbitMq\RabbitMqPublishingHelper;
 use Shlinkio\Shlink\IpGeolocation\GeoLite2\DbUpdater;
 use Shlinkio\Shlink\IpGeolocation\Resolver\IpLocationResolverInterface;
 use Symfony\Component\Mercure\Hub;
@@ -22,10 +22,14 @@ return [
         ],
         'async' => [
             EventDispatcher\Event\VisitLocated::class => [
-                EventDispatcher\NotifyVisitToMercure::class,
-                EventDispatcher\NotifyVisitToRabbitMq::class,
+                EventDispatcher\Mercure\NotifyVisitToMercure::class,
+                EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class,
                 EventDispatcher\NotifyVisitToWebHooks::class,
                 EventDispatcher\UpdateGeoLiteDb::class,
+            ],
+            EventDispatcher\Event\ShortUrlCreated::class => [
+                EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class,
+                EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class,
             ],
         ],
     ],
@@ -34,16 +38,24 @@ return [
         'factories' => [
             EventDispatcher\LocateVisit::class => ConfigAbstractFactory::class,
             EventDispatcher\NotifyVisitToWebHooks::class => ConfigAbstractFactory::class,
-            EventDispatcher\NotifyVisitToMercure::class => ConfigAbstractFactory::class,
-            EventDispatcher\NotifyVisitToRabbitMq::class => ConfigAbstractFactory::class,
+            EventDispatcher\Mercure\NotifyVisitToMercure::class => ConfigAbstractFactory::class,
+            EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class => ConfigAbstractFactory::class,
+            EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class => ConfigAbstractFactory::class,
+            EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => ConfigAbstractFactory::class,
             EventDispatcher\UpdateGeoLiteDb::class => ConfigAbstractFactory::class,
         ],
 
         'delegators' => [
-            EventDispatcher\NotifyVisitToMercure::class => [
+            EventDispatcher\Mercure\NotifyVisitToMercure::class => [
                 EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
             ],
-            EventDispatcher\NotifyVisitToRabbitMq::class => [
+            EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class => [
+                EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
+            ],
+            EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class => [
+                EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
+            ],
+            EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => [
                 EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
             ],
             EventDispatcher\NotifyVisitToWebHooks::class => [
@@ -68,17 +80,31 @@ return [
             ShortUrl\Transformer\ShortUrlDataTransformer::class,
             Options\AppOptions::class,
         ],
-        EventDispatcher\NotifyVisitToMercure::class => [
+        EventDispatcher\Mercure\NotifyVisitToMercure::class => [
             Hub::class,
             Mercure\MercureUpdatesGenerator::class,
             'em',
             'Logger_Shlink',
         ],
-        EventDispatcher\NotifyVisitToRabbitMq::class => [
-            AMQPStreamConnection::class,
+        EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class => [
+            Hub::class,
+            Mercure\MercureUpdatesGenerator::class,
+            'em',
+            'Logger_Shlink',
+        ],
+        EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class => [
+            RabbitMqPublishingHelper::class,
             'em',
             'Logger_Shlink',
             Visit\Transformer\OrphanVisitDataTransformer::class,
+            ShortUrl\Transformer\ShortUrlDataTransformer::class,
+            'config.rabbitmq.enabled',
+        ],
+        EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => [
+            RabbitMqPublishingHelper::class,
+            'em',
+            'Logger_Shlink',
+            ShortUrl\Transformer\ShortUrlDataTransformer::class,
             'config.rabbitmq.enabled',
         ],
         EventDispatcher\UpdateGeoLiteDb::class => [GeolocationDbUpdater::class, 'Logger_Shlink'],