+++ module/Core/src/Mercure/MercureUpdatesGenerator.php
@@ -5,26 +5,24 @@ declare(strict_types=1);
 namespace Shlinkio\Shlink\Core\Mercure;
 
 use Shlinkio\Shlink\Common\Rest\DataTransformerInterface;
+use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Entity\Visit;
+use Shlinkio\Shlink\Core\EventDispatcher\Topic;
 use Symfony\Component\Mercure\Update;
 
 use function Shlinkio\Shlink\Common\json_encode;
-use function sprintf;
 
 final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
 {
-    private const NEW_VISIT_TOPIC = 'https://shlink.io/new-visit';
-    private const NEW_ORPHAN_VISIT_TOPIC = 'https://shlink.io/new-orphan-visit';
-
     public function __construct(
-        private DataTransformerInterface $shortUrlTransformer,
-        private DataTransformerInterface $orphanVisitTransformer,
+        private readonly DataTransformerInterface $shortUrlTransformer,
+        private readonly DataTransformerInterface $orphanVisitTransformer,
     ) {
     }
 
     public function newVisitUpdate(Visit $visit): Update
     {
-        return new Update(self::NEW_VISIT_TOPIC, json_encode([
+        return new Update(Topic::NEW_VISIT->value, json_encode([
             'shortUrl' => $this->shortUrlTransformer->transform($visit->getShortUrl()),
             'visit' => $visit,
         ]));
@@ -32,7 +30,7 @@ final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
 
     public function newOrphanVisitUpdate(Visit $visit): Update
     {
-        return new Update(self::NEW_ORPHAN_VISIT_TOPIC, json_encode([
+        return new Update(Topic::NEW_ORPHAN_VISIT->value, json_encode([
             'visit' => $this->orphanVisitTransformer->transform($visit),
         ]));
     }
@@ -40,11 +38,18 @@ final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
     public function newShortUrlVisitUpdate(Visit $visit): Update
     {
         $shortUrl = $visit->getShortUrl();
-        $topic = sprintf('%s/%s', self::NEW_VISIT_TOPIC, $shortUrl?->getShortCode());
+        $topic = Topic::newShortUrlVisit($shortUrl?->getShortCode());
 
         return new Update($topic, json_encode([
             'shortUrl' => $this->shortUrlTransformer->transform($shortUrl),
             'visit' => $visit,
+        ]));
+    }
+
+    public function newShortUrlUpdate(ShortUrl $shortUrl): Update
+    {
+        return new Update(Topic::NEW_SHORT_URL->value, json_encode([
+            'shortUrl' => $this->shortUrlTransformer->transform($shortUrl),
         ]));
     }
 }