+++ module/Core/src/Mercure/MercureUpdatesGeneratorInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core\Mercure;
 
+use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Entity\Visit;
 use Symfony\Component\Mercure\Update;
 
@@ -14,4 +15,6 @@ interface MercureUpdatesGeneratorInterface
     public function newOrphanVisitUpdate(Visit $visit): Update;
 
     public function newShortUrlVisitUpdate(Visit $visit): Update;
+
+    public function newShortUrlUpdate(ShortUrl $shortUrl): Update;
 }