+++ module/Core/test/EventDispatcher/RabbitMq/NotifyVisitToRabbitMqTest.php
@@ -2,25 +2,26 @@
 
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\Core\EventDispatcher;
+namespace ShlinkioTest\Shlink\Core\EventDispatcher\RabbitMq;
 
 use Doctrine\ORM\EntityManagerInterface;
 use DomainException;
 use Exception;
-use PhpAmqpLib\Channel\AMQPChannel;
-use PhpAmqpLib\Connection\AMQPStreamConnection;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
+use Shlinkio\Shlink\Common\RabbitMq\RabbitMqPublishingHelperInterface;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Entity\Visit;
 use Shlinkio\Shlink\Core\EventDispatcher\Event\VisitLocated;
-use Shlinkio\Shlink\Core\EventDispatcher\NotifyVisitToRabbitMq;
+use Shlinkio\Shlink\Core\EventDispatcher\RabbitMq\NotifyVisitToRabbitMq;
 use Shlinkio\Shlink\Core\Model\ShortUrlMeta;
 use Shlinkio\Shlink\Core\Model\Visitor;
+use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortUrlStringifier;
+use Shlinkio\Shlink\Core\ShortUrl\Transformer\ShortUrlDataTransformer;
 use Shlinkio\Shlink\Core\Visit\Transformer\OrphanVisitDataTransformer;
 use Throwable;
 
@@ -32,28 +33,22 @@ class NotifyVisitToRabbitMqTest extends TestCase
     use ProphecyTrait;
 
     private NotifyVisitToRabbitMq $listener;
-    private ObjectProphecy $connection;
+    private ObjectProphecy $helper;
     private ObjectProphecy $em;
     private ObjectProphecy $logger;
-    private ObjectProphecy $orphanVisitTransformer;
-    private ObjectProphecy $channel;
 
     protected function setUp(): void
     {
-        $this->channel = $this->prophesize(AMQPChannel::class);
-
-        $this->connection = $this->prophesize(AMQPStreamConnection::class);
-        $this->connection->isConnected()->willReturn(false);
-        $this->connection->channel()->willReturn($this->channel->reveal());
-
+        $this->helper = $this->prophesize(RabbitMqPublishingHelperInterface::class);
         $this->em = $this->prophesize(EntityManagerInterface::class);
         $this->logger = $this->prophesize(LoggerInterface::class);
 
         $this->listener = new NotifyVisitToRabbitMq(
-            $this->connection->reveal(),
+            $this->helper->reveal(),
             $this->em->reveal(),
             $this->logger->reveal(),
             new OrphanVisitDataTransformer(),
+            new ShortUrlDataTransformer(new ShortUrlStringifier([])),
             true,
         );
     }
@@ -62,10 +57,11 @@ class NotifyVisitToRabbitMqTest extends TestCase
     public function doesNothingWhenTheFeatureIsNotEnabled(): void
     {
         $listener = new NotifyVisitToRabbitMq(
-            $this->connection->reveal(),
+            $this->helper->reveal(),
             $this->em->reveal(),
             $this->logger->reveal(),
             new OrphanVisitDataTransformer(),
+            new ShortUrlDataTransformer(new ShortUrlStringifier([])),
             false,
         );
 
@@ -74,8 +70,7 @@ class NotifyVisitToRabbitMqTest extends TestCase
         $this->em->find(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->logger->warning(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->logger->debug(Argument::cetera())->shouldNotHaveBeenCalled();
-        $this->connection->isConnected()->shouldNotHaveBeenCalled();
-        $this->connection->close()->shouldNotHaveBeenCalled();
+        $this->helper->publishPayloadInQueue(Argument::cetera())->shouldNotHaveBeenCalled();
     }
 
     /** @test */
@@ -93,8 +88,7 @@ class NotifyVisitToRabbitMqTest extends TestCase
         $findVisit->shouldHaveBeenCalledOnce();
         $logWarning->shouldHaveBeenCalledOnce();
         $this->logger->debug(Argument::cetera())->shouldNotHaveBeenCalled();
-        $this->connection->isConnected()->shouldNotHaveBeenCalled();
-        $this->connection->close()->shouldNotHaveBeenCalled();
+        $this->helper->publishPayloadInQueue(Argument::cetera())->shouldNotHaveBeenCalled();
     }
 
     /**
@@ -105,27 +99,17 @@ class NotifyVisitToRabbitMqTest extends TestCase
     {
         $visitId = '123';
         $findVisit = $this->em->find(Visit::class, $visitId)->willReturn($visit);
-        $argumentWithExpectedChannel = Argument::that(fn (string $channel) => contains($expectedChannels, $channel));
+        $argumentWithExpectedChannels = Argument::that(
+            static fn (string $channel) => contains($expectedChannels, $channel),
+        );
 
         ($this->listener)(new VisitLocated($visitId));
 
         $findVisit->shouldHaveBeenCalledOnce();
-        $this->channel->exchange_declare($argumentWithExpectedChannel, Argument::cetera())->shouldHaveBeenCalledTimes(
-            count($expectedChannels),
-        );
-        $this->channel->queue_declare($argumentWithExpectedChannel, Argument::cetera())->shouldHaveBeenCalledTimes(
-            count($expectedChannels),
-        );
-        $this->channel->queue_bind(
-            $argumentWithExpectedChannel,
-            $argumentWithExpectedChannel,
+        $this->helper->publishPayloadInQueue(
+            Argument::type('array'),
+            $argumentWithExpectedChannels,
         )->shouldHaveBeenCalledTimes(count($expectedChannels));
-        $this->channel->basic_publish(Argument::any(), $argumentWithExpectedChannel)->shouldHaveBeenCalledTimes(
-            count($expectedChannels),
-        );
-        $this->channel->close()->shouldHaveBeenCalledOnce();
-        $this->connection->reconnect()->shouldHaveBeenCalledOnce();
-        $this->connection->close()->shouldHaveBeenCalledOnce();
         $this->logger->debug(Argument::cetera())->shouldNotHaveBeenCalled();
     }
 
@@ -154,7 +138,7 @@ class NotifyVisitToRabbitMqTest extends TestCase
     {
         $visitId = '123';
         $findVisit = $this->em->find(Visit::class, $visitId)->willReturn(Visit::forBasePath(Visitor::emptyInstance()));
-        $channel = $this->connection->channel()->willThrow($e);
+        $publish = $this->helper->publishPayloadInQueue(Argument::cetera())->willThrow($e);
 
         ($this->listener)(new VisitLocated($visitId));
 
@@ -162,11 +146,8 @@ class NotifyVisitToRabbitMqTest extends TestCase
             'Error while trying to notify RabbitMQ with new visit. {e}',
             ['e' => $e],
         )->shouldHaveBeenCalledOnce();
-        $this->connection->close()->shouldHaveBeenCalledOnce();
-        $this->connection->reconnect()->shouldHaveBeenCalledOnce();
         $findVisit->shouldHaveBeenCalledOnce();
-        $channel->shouldHaveBeenCalledOnce();
-        $this->channel->close()->shouldNotHaveBeenCalled();
+        $publish->shouldHaveBeenCalledOnce();
     }
 
     public function provideExceptions(): iterable